requirements.txt
functions-framework==3.*
google-api-python-client>=2.100.0
google-auth>=2.22.0
google-cloud-storage>=2.10.0
google-resumable-media>=2.7.0
requests>=2.28.0
ffmpeg-python>=0.2.0
